Updated README to reflect that this repository now accepts contributed JSON-Schema files in the "contrib" directory.
Enhanced the JsonSchemaTest suite to be able to run all the good/bad examples through one of the contributed JSON-Schema files easily -- by simply editing the path and uncommenting out the contributed tests (they are uncommented out in this checkin, as an example).

Added the test results for the latest contribution.